refactor: moved QuestionFactory interface to a proper place
