+++ review/src/main/kotlin/br/all/infrastructure/question/factory/NumberScaleQuestionFactory.kt
@@ -2,7 +2,7 @@ package br.all.infrastructure.question.factory
 
 import br.all.application.question.create.CreateQuestionService
 import br.all.application.question.repository.QuestionDto
-import br.all.application.question.repository.QuestionRepository
+import br.all.application.question.shared.QuestionFactory
 import br.all.domain.model.protocol.ProtocolId
 import br.all.domain.model.question.NumberScale
 import br.all.domain.model.question.Question
@@ -12,7 +12,7 @@ import br.all.infrastructure.question.MongoQuestionRepository
 import br.all.infrastructure.question.QuestionRepositoryImpl
 import java.util.*
 
-class NumberScaleQuestionFactory(private val mongoRepository: MongoQuestionRepository): QuestionFactory{
+class NumberScaleQuestionFactory(private val mongoRepository: MongoQuestionRepository): QuestionFactory {
     override fun create(id: UUID, request: CreateQuestionService.RequestModel) = with(request){
         if (higher == null || lower == null) throw IllegalArgumentException("Higher and Lower values must not be null.")
         QuestionBuilder.with(QuestionId(id), ProtocolId(protocolId), code, description)
