+++ review/src/main/kotlin/br/all/infrastructure/question/factory/TextualQuestionFactory.kt
@@ -2,6 +2,7 @@ package br.all.infrastructure.question.factory
 
 import br.all.application.question.create.CreateQuestionService.RequestModel
 import br.all.application.question.repository.QuestionDto
+import br.all.application.question.shared.QuestionFactory
 import br.all.domain.model.protocol.ProtocolId
 import br.all.domain.model.question.Question
 import br.all.domain.model.question.QuestionBuilder
